// src/app/views/model/card.model.ts
export class CardModel {
	id: string;
	customerId: string;
	cardName: string;
	cardRecordID: string;
	cardType: string;
	promotionId: string;
	transProId: string;
	ycomer: string;
	memberCode: string;
	memberDateExpored?: any;
	infoList: string;
	infoDetail: string;
	status: string;
	timeUpdate: Date;
	inputter: string;
	authoriser: string;
	timeAuth?: any;
	coCode: string;
	cardCheckbox: boolean;
	accountNumber: any;
	actionT24?: string;
	homeAddress?: string;
	cardNumber?: string;
}