// src/app/views/model/ebank.model.ts
export class EbankModel {
	id: string;
	serviceId: string;
	status: string;
	timeUpdate: string;
	inputter: string;
	authoriser: string;
	timeAuth: string;
	coCode: string;
	customerId: string;
	username: string;
	packageService: string;
	passwordType: string;
	promotionId: string;
	limit: string;
	transSeANetPro: string;
	account: string;
	ebankCheckbox: boolean;
}